Move MappingSetter to mapping package
